1. 进货增加字段 2. 调整数据库数据字段类型

+++ flaskr/models/batch_out.py
@@ -5,9 +5,9 @@ from marshmallow import Schema, fields, validate
 # 出库批次表
 class Batch_out(db.Model):
   id          = db.Column(db.Integer, primary_key=True)
-  out_number  = db.Column(db.Float(6,1), nullable=False, comment='出库数量')
-  price       = db.Column(db.Float(6,1), nullable=False, comment='出售单价')
-  loss        = db.Column(db.Float(6,1), default=0, comment='损耗')
+  out_number  = db.Column(db.DECIMAL(6,1), nullable=False, comment='出库数量')
+  price       = db.Column(db.DECIMAL(6,1), nullable=False, comment='出售单价')
+  loss        = db.Column(db.DECIMAL(6,1), default=0, comment='损耗')
   document    = db.Column(db.String(20), comment='单据编号')
   code        = db.Column(db.String(20), comment='物料长代码')
   comment     = db.Column(db.Text, comment='备注')
@@ -20,10 +20,10 @@ class Batch_out(db.Model):
 
 class Batch_out_schema(Schema):
   id          = fields.Integer(dump_only=True)
-  out_number  = fields.Float()
+  out_number  = fields.Decimal()
   out_time    = fields.DateTime()
-  price       = fields.Float()
-  loss        = fields.Float()
+  price       = fields.Decimal()
+  loss        = fields.Decimal()
   document    = fields.String()
   code        = fields.String()
   comment     = fields.String()
